admin: add FormModelBase class
This will be the root class for form models, holding onto things that
aren't specific to any particular form (e.g. validation). This also
changes LiveItemFormModel to inherit from FormModelBase, with the
generic validation code moved to the latter.
